+++ history.md
@@ -2,6 +2,11 @@
 
 v1.1.0dev
 - Add citation information.
+- Add status bar messages for different UI components.
+- Enhance status bar updates by a signal-based mechanism.
+- Fix: remove old clicked points, so they do not reappear after changing CSR.
+- Check the selected layer and deactivate plugin if the layer is not compatible with InSAR Explorer.
+- Add support for polygon selection for vector data.
 - Change the reference point for the map on the fly.
 - Enhance labeling of the map plot
 - Enhance curve fitting by a prior normalizing of dates.